moved add_counts_to_yticklabels to utils.py plus a test
